Remove code duplication by moving inner classes to outer level
